+++ src/App.css
@@ -1,0 +1,3 @@
+*{
+    margin: 0px;
+}
