+++ src/index.js
@@ -1,16 +1,12 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './mycomponents/App.css';
-import App from './mycomponents/App';
-import { BrowserRouter,Switch } from 'react-router-dom';
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "./App.css";
+import App from "./App";
+import { BrowserRouter } from "react-router-dom";
 
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
+const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-<BrowserRouter>
+  <BrowserRouter>
     <App />
-
-    </BrowserRouter>
-
+  </BrowserRouter>
 );
-
