--- src/mycomponents/App.css
@@ -1,3 +1,0 @@
-*{
-    margin: 0px;
-}
