--- src/mycomponents/App.js
@@ -1,20 +1,0 @@
-import React from "react";
-import Header from "./Header";
-import Main from "./Main";
-import Home from "./Home";
-import { Route,Routes } from 'react-router-dom';
-
-const App = () => {
-  return (
-<>
- <Routes>
-<Route path="/" element={<Home />} />
-<Route path="/Home" element={<Home/>} />  
-<Route path="/Main" element={<Main/>} />  
-</Routes>
- 
-  
-    </>
-  );
-};
-export default App;
